training/requirements.txt
fastapi
uvicorn
python-multipart
numpy
torch
Pillow
opencv-python
tqdm
torchvision
pyyaml
requests
matplotlib
seaborn
scipy
pandas
scikit-learn
aiofiles